# src/main/webui/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "ConfirmImportRequest" : {
        "type" : "object",
        "properties" : {
          "recipes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ConfirmedRecipe"
            }
          }
        }
      },
      "ConfirmedRecipe" : {
        "type" : "object",
        "properties" : {
          "recipeName" : {
            "type" : "string"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ingredient" : {
            "type" : "string"
          },
          "keep" : {
            "type" : "boolean"
          }
        }
      },
      "CreateCookbookRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "author" : {
            "type" : "string"
          }
        }
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    }
  },
  "paths" : {
    "/api/cookbooks" : {
      "get" : {
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "List Cookbooks",
        "tags" : [ "Cookbook Resource" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCookbookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Cookbook",
        "tags" : [ "Cookbook Resource" ]
      }
    },
    "/api/cookbooks/{id}" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Cookbook",
        "tags" : [ "Cookbook Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Delete Cookbook",
        "tags" : [ "Cookbook Resource" ]
      }
    },
    "/api/cookbooks/{id}/confirm" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmImportRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Confirm Import",
        "tags" : [ "Cookbook Resource" ]
      }
    },
    "/api/cookbooks/{id}/cover" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Upload Cover",
        "tags" : [ "Cookbook Resource" ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "image/jpeg" : { },
              "image/png" : { }
            }
          }
        },
        "summary" : "Get Cover",
        "tags" : [ "Cookbook Resource" ]
      }
    },
    "/api/cookbooks/{id}/index-pages" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "files" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Upload Index Pages",
        "tags" : [ "Cookbook Resource" ]
      }
    },
    "/api/cookbooks/{id}/ocr/start" : {
      "post" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Start Ocr Processing",
        "tags" : [ "Cookbook Resource" ]
      }
    },
    "/api/cookbooks/{id}/recipes" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get Cookbook Recipes",
        "tags" : [ "Cookbook Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "the-recipe-db API",
    "version" : "1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8081",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8081",
    "description" : "Auto generated value"
  } ]
}